+++ packages/perspective-viewer-d3fc/src/js/series/seriesRange.js
@@ -7,7 +7,7 @@
  *
  */
 import * as d3 from "d3";
-import {domain} from "../axis/linear";
+import {domain} from "../axis/linearAxis";
 
 export function seriesLinearRange(settings, data, valueName) {
     return d3.scaleLinear().domain(
